[OBS] renamed SetMute to ToggleMute
