+++ obssuite/togglemute/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.stream-pi</groupId>
-    <artifactId>obssuite_setmuteaction</artifactId>
+    <artifactId>obssuite_togglemuteaction</artifactId>
     <version>2.0.0</version>
 
     <build>
