+++ obssuite/togglemute/src/main/java/module-info.java
@@ -1,0 +1,12 @@
+import togglemute.ToggleMute;
+
+module com.stream_pi.obssuite.togglemuteaction
+{
+    requires com.stream_pi.action_api;
+    requires com.stream_pi.util;
+
+    requires obs.websocket.java;
+    requires com.stream_pi.obssuite.motheraction;
+
+    provides com.stream_pi.action_api.externalplugin.ExternalPlugin with ToggleMute;
+}
